pyproject.toml
[project]
name = "ShadowCrypt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "filelock>=3.18.0",
    "pycryptodome>=3.22.0",
    "pyinstaller>=6.12.0",
    "pylnk3>=0.4.2",
]
